requirements.txt
greenlet==1.1.3.post0
SQLAlchemy==1.4.42
PyYAML==6.0
Jinja2~=3.1.2
PyMySQL==1.0.2